support for storing some fields of  SampleInfo
